test: check the row count displayed in grid page
