+++ frontend/appflowy_flutter/integration_test/database_field_test.dart
@@ -3,6 +3,7 @@ import 'package:appflowy_backend/protobuf/flowy-database2/field_entities.pbenum.
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
+import 'util/database_test_op.dart';
 import 'util/util.dart';
 
 void main() {
@@ -93,51 +94,6 @@ void main() {
         await tester.findCellByFieldType(fieldType);
         await tester.pumpAndSettle();
       }
-    });
-
-    testWidgets('create row of the grid', (tester) async {
-      await tester.initializeAppFlowy();
-      await tester.tapGoButton();
-
-      await tester.tapAddButton();
-      await tester.tapCreateGridButton();
-      await tester.tapCreateRowButtonInGrid();
-
-      // The initial number of rows is 3
-      await tester.assertNumberOfRowsInGridPage(4);
-      await tester.pumpAndSettle();
-    });
-
-    testWidgets('create row from row menu of the grid', (tester) async {
-      await tester.initializeAppFlowy();
-      await tester.tapGoButton();
-
-      await tester.tapAddButton();
-      await tester.tapCreateGridButton();
-      await tester.hoverOnFirstRowOfGrid();
-
-      await tester.tapCreateRowButtonInRowMenuOfGrid();
-
-      // The initial number of rows is 3
-      await tester.assertNumberOfRowsInGridPage(4);
-      await tester.pumpAndSettle();
-    });
-
-    testWidgets('delete row of the grid', (tester) async {
-      await tester.initializeAppFlowy();
-      await tester.tapGoButton();
-
-      await tester.tapAddButton();
-      await tester.tapCreateGridButton();
-      await tester.hoverOnFirstRowOfGrid();
-
-      // Open the row menu and then click the delete
-      await tester.tapRowMenuButtonInGrid();
-      await tester.tapDeleteOnRowMenu();
-
-      // The initial number of rows is 3
-      await tester.assertNumberOfRowsInGridPage(2);
-      await tester.pumpAndSettle();
     });
   });
 }
