+++ frontend/appflowy_flutter/integration_test/database_row_detail_page_test.dart
@@ -1,10 +1,10 @@
 import 'package:appflowy/plugins/database_view/widgets/row/row_banner.dart';
 import 'package:appflowy_backend/protobuf/flowy-database2/field_entities.pbenum.dart';
 import 'package:flowy_infra_ui/style_widget/text.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
+import 'util/database_test_op.dart';
 import 'util/util.dart';
 
 void main() {
